+++ sdk/nodejs/cloudkms/v1/ekmConnection.ts
@@ -7,7 +7,6 @@ import * as utilities from "../../utilities";
 
 /**
  * Creates a new EkmConnection in a given Project and Location.
- * Auto-naming is currently not supported for this resource.
  * Note - this resource's API doesn't support deletion. When deleted, the resource will persist
  * on Google Cloud even though it will be deleted from Pulumi state.
  */
@@ -62,13 +61,10 @@ export class EkmConnection extends pulumi.CustomResource {
      * @param args The arguments to use to populate this resource's properties.
      * @param opts A bag of options that control this resource's behavior.
      */
-    constructor(name: string, args: EkmConnectionArgs, opts?: pulumi.CustomResourceOptions) {
+    constructor(name: string, args?: EkmConnectionArgs, opts?: pulumi.CustomResourceOptions) {
         let resourceInputs: pulumi.Inputs = {};
         opts = opts || {};
         if (!opts.id) {
-            if ((!args || args.ekmConnectionId === undefined) && !opts.urn) {
-                throw new Error("Missing required property 'ekmConnectionId'");
-            }
             resourceInputs["ekmConnectionId"] = args ? args.ekmConnectionId : undefined;
             resourceInputs["etag"] = args ? args.etag : undefined;
             resourceInputs["location"] = args ? args.location : undefined;
@@ -91,7 +87,7 @@ export class EkmConnection extends pulumi.CustomResource {
  * The set of arguments for constructing a EkmConnection resource.
  */
 export interface EkmConnectionArgs {
-    ekmConnectionId: pulumi.Input<string>;
+    ekmConnectionId?: pulumi.Input<string>;
     /**
      * This checksum is computed by the server based on the value of other fields, and may be sent on update requests to ensure the client has an up-to-date value before proceeding.
      */